# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nutrition Assistant API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/meals": {
      "post": {
        "tags": [
          "meals"
        ],
        "summary": "Create Meal",
        "operationId": "create_meal_api_v1_meals_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Meal Api V1 Meals Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "meals"
        ],
        "summary": "List Meals",
        "operationId": "list_meals_api_v1_meals_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealListEnvelopeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meals/{meal_id}": {
      "get": {
        "tags": [
          "meals"
        ],
        "summary": "Get Meal",
        "operationId": "get_meal_api_v1_meals__meal_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "meal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealReadEnvelopeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "meals"
        ],
        "summary": "Update Meal",
        "operationId": "update_meal_api_v1_meals__meal_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "meal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Meal Api V1 Meals  Meal Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "meals"
        ],
        "summary": "Delete Meal",
        "operationId": "delete_meal_api_v1_meals__meal_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "meal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Meal Api V1 Meals  Meal Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meals/analysis": {
      "post": {
        "tags": [
          "meals"
        ],
        "summary": "Analyze Meal",
        "operationId": "analyze_meal_api_v1_meals_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Analyze Meal Api V1 Meals Analysis Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/meal-plans": {
      "get": {
        "tags": [
          "meal-plans"
        ],
        "summary": "List Plans",
        "operationId": "list_plans_api_v1_meal_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Plans Api V1 Meal Plans Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "meal-plans"
        ],
        "summary": "Create Plan",
        "operationId": "create_plan_api_v1_meal_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealPlanCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Plan Api V1 Meal Plans Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/meal-plans/{plan_id}": {
      "patch": {
        "tags": [
          "meal-plans"
        ],
        "summary": "Update Plan",
        "operationId": "update_plan_api_v1_meal_plans__plan_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealPlanUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Plan Api V1 Meal Plans  Plan Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "meal-plans"
        ],
        "summary": "Delete Plan",
        "operationId": "delete_plan_api_v1_meal_plans__plan_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Plan Api V1 Meal Plans  Plan Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MealAnalysisRequest": {
        "properties": {
          "meal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meal Id"
          },
          "uploaded_image_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Uploaded Image Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "MealAnalysisRequest"
      },
      "MealCreateRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "meal_type": {
            "$ref": "#/components/schemas/MealType"
          },
          "source": {
            "$ref": "#/components/schemas/MealSource",
            "default": "manual"
          },
          "eaten_at": {
            "type": "string",
            "format": "date-time",
            "title": "Eaten At"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MealItemCreateRequest"
            },
            "type": "array",
            "minItems": 1,
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meal_type",
          "eaten_at"
        ],
        "title": "MealCreateRequest"
      },
...
      "MealItemCreateRequest": {
        "properties": {
          "ingredient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredient Id"
          },
          "food_product_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Food Product Id"
          },
          "display_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Display Name"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Unit",
            "default": "g"
          },
          "position": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Position",
            "default": 0
          },
          "nutrition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NutritionValueDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "display_name",
          "quantity"
        ],
        "title": "MealItemCreateRequest"
      },
...
      "MealListDataResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MealReadResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "MealListDataResponse"
      },
      "MealListEnvelopeResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "message_key": {
            "type": "string",
            "title": "Message Key"
          },
          "data": {
            "$ref": "#/components/schemas/MealListDataResponse"
          },
          "error": {
            "type": "null",
            "title": "Error"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message_key",
          "data"
        ],
        "title": "MealListEnvelopeResponse"
      },
...
      "MealPlanCreateRequest": {
        "properties": {
          "plan_date": {
            "type": "string",
            "format": "date-time",
            "title": "Plan Date"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "$ref": "#/components/schemas/MealPlanStatus",
            "default": "draft"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/MealPlanEntryDTO"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan_date"
        ],
        "title": "MealPlanCreateRequest"
      },
      "MealPlanEntryDTO": {
        "properties": {
          "meal_type": {
            "$ref": "#/components/schemas/MealType"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meal_type",
          "title"
        ],
        "title": "MealPlanEntryDTO"
      },
      "MealPlanStatus": {
        "type": "string",
        "enum": [
          "draft",
          "active",
          "archived"
        ],
        "title": "MealPlanStatus"
      },
      "MealPlanUpdateRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MealPlanStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "entries": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MealPlanEntryDTO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entries"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "MealPlanUpdateRequest"
      },
      "MealReadEnvelopeResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "message_key": {
            "type": "string",
            "title": "Message Key"
          },
          "data": {
            "$ref": "#/components/schemas/MealReadResponse"
          },
          "error": {
            "type": "null",
            "title": "Error"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message_key",
          "data"
        ],
        "title": "MealReadEnvelopeResponse"
      },
...
      "MealSource": {
        "type": "string",
        "enum": [
          "manual",
          "barcode",
          "image_analysis",
          "imported"
        ],
        "title": "MealSource"
      },
      "MealType": {
        "type": "string",
        "enum": [
          "breakfast",
          "lunch",
          "dinner",
          "snack",
          "other"
        ],
        "title": "MealType"
      },
      "MealUpdateRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "meal_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MealType"
              },
              {
                "type": "null"
              }
            ]
          },
          "eaten_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eaten At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "MealUpdateRequest"
      },
...
      "PaginationMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "total_pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Pages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "title": "PaginationMeta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}